--- a/helpers_fuzz_test.go
+++ b/helpers_fuzz_test.go
@@ -7,7 +7,7 @@ import (
 )
 
 // go test -v -run=^$ -fuzz=Fuzz_Utils_GetOffer
-func Fuzz_Utils_GetOffer(f *testing.F) {
+func FuzzUtilsGetOffer(f *testing.F) {
 	bigHeader := `application/json; v=1; foo=bar; q=0.938; extra=param, text/plain;param="big fox"; q=0.43`
 	f.Add(bigHeader)
 	f.Fuzz(func(_ *testing.T, spec string) {