Attempts to approach the level of support offered by express, but behavior may differ in unusual corner cases.
Some key behaviors from Express that are implemented:
 - If an offer does not have every parameter listed in the given Accept, it is rejected.
 - Parameters do not affect specificity.
 - In a given specificity, more parameters gives greater precedence
 - Parameters are unordered
 - Matching is case-insensitive
 - Surrounding quotes behave strangely (buggy?) in Express, so we only explicitly handle what RFC-9110 allows, which is a quoted value (right side of the equal sign)

Behaviors that were intentionally **not** implemented:
- RFC 9110 defines quoted-pairs as octet pairs consisting of a backslash followed by another character. It states that recipients (i.e., servers) must parse quoted pairs as if they were replaced by the octet following the backslash. This changeset does not "ignore" the backslash of quoted-pairs.
Justification: this only affects functionality when a client escapes a character other than " or \ _and_ the server doesn't account for it in the call to `ctx.Accepts`. This case is unusual, and RFC 9110 states that clients ought not to do this. See doc comment for paramsMatch for more details.
- The following case is not handled yet: `Accept: */*; param=foo, text/plain; q=0.9, text/plain; param=foo; q=0.8`. If the server accepts `text/plain` and `text/plain;param=foo`, Express will match "text/plain" because the client explicitly stated they prefer `text/plain` over `text/plain;param=foo` _even though_ `*/*;param=foo` has quality 1.0.
Justification: this behavior was already present in the existing code.

Behaviors were mostly reverse engineered from res.format in Express.

https://www.rfc-editor.org/rfc/rfc9110#name-parameters

## More Detail

We redundantly parse parameters for each offer N times instead of just once. Parsing them once in getOffer and passing the results to isAccepted indeed speeds things up, but it slows down the "normal" case (no parameters) more than I would like.

## Type of change

- [x] New feature (non-breaking change which adds functionality)
- [x] This change requires a documentation update

## Checklist:

- [x] For new functionalities I follow the inspiration of the express js framework and built them similar in usage
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation - /docs/ directory for https://docs.gofiber.io/
- [x] I have added tests that prove my fix is effective or that my feature works
- [x] New and existing unit tests pass locally with my changes
- [x] I tried to make my code as fast as possible with as few allocations as possible
- [x] For new code I have written benchmarks so that they can be analyzed and improved