@nickajacks1 2023-10-15
Ah, I see the CI includes go 1.17, but fuzzing was introduced in 1.18.
Should I remove the fuzz test/move it to a different file with a build constraint so it only runs on 1.18+?