@nickajacks1 2023-10-16
Another thing to note is that this does NOT support [RFC 2231](https://datatracker.ietf.org/doc/html/rfc2231), which the stdlib mime package uses for ParseMediaType.  mime.ParseMediaType is comparativesly slow and results in more allocations, but might simplify the code significantly (and give fiber RFC 2231 support).  To give perspective, it's about 4-5 times slower than getHeaderValParams on my PC.  On the other hand,  it's on the order of about a microsecond or 2 difference, so it might be worth it.  Dunno.

Express's res.format doesn't seem to support it (I did only a very quick test).