@sixcolors 2023-10-16
Express Accepts is very slow. See https://github.com/gofiber/fiber/issues/2387 for some discussion of Accepts and getOffer https://github.com/gofiber/fiber/pull/2486