@nickajacks1 2023-11-05
### Bench Accepts

<details>
<summary>master</summary>

```
goos: linux
goarch: amd64
pkg: github.com/gofiber/fiber/v2
cpu: 11th Gen Intel(R) Core(TM) i7-1165G7 @ 2.80GHz
Benchmark_Ctx_Accepts/run-[]string{".xml"}-8         	 8839486	       135.0 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{".xml"}-8         	 8853195	       139.1 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{".xml"}-8         	 7805179	       135.1 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{".xml"}-8         	 8777102	       140.6 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{".xml"}-8         	 8532048	       135.7 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{".xml"}-8         	 8545423	       136.7 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{".xml"}-8         	 8990457	       134.5 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{".xml"}-8         	 8810748	       135.9 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{".xml"}-8         	 8968890	       136.1 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{".xml"}-8         	 8821316	       135.8 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{".xml"}-8         	 8952948	       136.1 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{".xml"}-8         	 8813887	       135.9 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{".xml"}-8         	 8967898	       135.2 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{".xml"}-8         	 8391528	       134.0 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{".xml"}-8         	 8741970	       143.2 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"json",_"xml"}-8  	 6189565	       188.6 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"json",_"xml"}-8  	 6358563	       190.5 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"json",_"xml"}-8  	 6245946	       196.7 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"json",_"xml"}-8  	 6351415	       186.3 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"json",_"xml"}-8  	 6215264	       187.0 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"json",_"xml"}-8  	 6279278	       186.6 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"json",_"xml"}-8  	 6415616	       189.7 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"json",_"xml"}-8  	 6171139	       198.6 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"json",_"xml"}-8  	 6187696	       192.0 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"json",_"xml"}-8  	 6549969	       196.3 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"json",_"xml"}-8  	 6304629	       186.6 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"json",_"xml"}-8  	 6336711	       189.2 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"json",_"xml"}-8  	 6361267	       192.1 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"json",_"xml"}-8  	 6310435	       183.0 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"json",_"xml"}-8  	 6247321	       194.6 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"application/json",_"application/xml"}-8         	 8753716	       139.9 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"application/json",_"application/xml"}-8         	 8693688	       140.7 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"application/json",_"application/xml"}-8         	 8493938	       141.0 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"application/json",_"application/xml"}-8         	 8311784	       146.2 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"application/json",_"application/xml"}-8         	 8706558	       147.6 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"application/json",_"application/xml"}-8         	 8805723	       141.4 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"application/json",_"application/xml"}-8         	 8746408	       138.7 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"application/json",_"application/xml"}-8         	 8558030	       140.4 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"application/json",_"application/xml"}-8         	 8614651	       146.0 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"application/json",_"application/xml"}-8         	 8621936	       146.6 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"application/json",_"application/xml"}-8         	 8708131	       142.0 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"application/json",_"application/xml"}-8         	 8589172	       144.3 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"application/json",_"application/xml"}-8         	 8213953	       146.1 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"application/json",_"application/xml"}-8         	 8415702	       146.2 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"application/json",_"application/xml"}-8         	 8660049	       147.3 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsCharsets-8                                                     	16403662	        71.86 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsCharsets-8                                                     	16912604	        72.46 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsCharsets-8                                                     	16084365	        70.73 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsCharsets-8                                                     	16696166	        69.75 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsCharsets-8                                                     	16815865	        71.65 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsCharsets-8                                                     	17089466	        73.08 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsCharsets-8                                                     	15634952	        70.18 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsCharsets-8                                                     	16885665	        72.69 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsCharsets-8                                                     	16786174	        73.04 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsCharsets-8                                                     	16277292	        69.21 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsCharsets-8                                                     	16778400	        69.51 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsCharsets-8                                                     	17192934	        74.30 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsCharsets-8                                                     	16519195	        69.91 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsCharsets-8                                                     	17105088	        70.72 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsCharsets-8                                                     	16805516	        69.26 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsEncodings-8                                                    	11841327	       101.7 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsEncodings-8                                                    	11701485	       109.4 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsEncodings-8                                                    	11197510	       100.4 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsEncodings-8                                                    	11380326	       105.1 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsEncodings-8                                                    	11736454	       104.1 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsEncodings-8                                                    	11458483	       102.8 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsEncodings-8                                                    	11492564	       104.1 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsEncodings-8                                                    	11742571	       106.1 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsEncodings-8                                                    	11194947	       109.4 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsEncodings-8                                                    	11642266	       100.3 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsEncodings-8                                                    	11775794	       105.2 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsEncodings-8                                                    	11217306	       107.9 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsEncodings-8                                                    	11892580	       105.5 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsEncodings-8                                                    	11464521	       109.9 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsEncodings-8                                                    	11852080	       106.5 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsLanguages-8                                                    	 7691950	       162.9 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsLanguages-8                                                    	 7349106	       157.9 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsLanguages-8                                                    	 7186768	       160.9 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsLanguages-8                                                    	 7588566	       153.4 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsLanguages-8                                                    	 7647355	       155.4 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsLanguages-8                                                    	 7767267	       153.8 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsLanguages-8                                                    	 7678732	       153.5 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsLanguages-8                                                    	 7717558	       158.7 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsLanguages-8                                                    	 7680273	       164.8 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsLanguages-8                                                    	 7745534	       153.3 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsLanguages-8                                                    	 7738284	       154.4 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsLanguages-8                                                    	 7577661	       152.7 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsLanguages-8                                                    	 7616619	       154.3 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsLanguages-8                                                    	 7461470	       154.4 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsLanguages-8                                                    	 7244853	       157.1 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/gofiber/fiber/v2	120.871s
```

</details>

<details>
<summary>PR</summary>

```
goos: linux
goarch: amd64
pkg: github.com/gofiber/fiber/v2
cpu: 11th Gen Intel(R) Core(TM) i7-1165G7 @ 2.80GHz
Benchmark_Ctx_Accepts/run-[]string{".xml"}-8         	 7096329	       164.1 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{".xml"}-8         	 7097638	       165.3 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{".xml"}-8         	 7407602	       166.4 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{".xml"}-8         	 7600789	       167.0 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{".xml"}-8         	 7393784	       165.4 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{".xml"}-8         	 7291954	       166.8 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{".xml"}-8         	 7315314	       164.6 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{".xml"}-8         	 7331866	       164.0 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{".xml"}-8         	 7442922	       166.2 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{".xml"}-8         	 7344050	       164.9 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{".xml"}-8         	 7211116	       171.4 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{".xml"}-8         	 7264797	       165.7 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{".xml"}-8         	 7486377	       168.6 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{".xml"}-8         	 7171075	       163.0 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{".xml"}-8         	 7379565	       161.5 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"json",_"xml"}-8  	 5563500	       223.8 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"json",_"xml"}-8  	 4985742	       223.5 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"json",_"xml"}-8  	 5433070	       216.4 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"json",_"xml"}-8  	 5434917	       220.5 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"json",_"xml"}-8  	 5531926	       225.4 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"json",_"xml"}-8  	 5486791	       222.3 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"json",_"xml"}-8  	 5542166	       224.6 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"json",_"xml"}-8  	 5271254	       223.1 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"json",_"xml"}-8  	 5591925	       216.3 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"json",_"xml"}-8  	 5265901	       226.3 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"json",_"xml"}-8  	 5612340	       215.5 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"json",_"xml"}-8  	 5348716	       227.8 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"json",_"xml"}-8  	 5654978	       215.0 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"json",_"xml"}-8  	 5468170	       220.0 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"json",_"xml"}-8  	 5377267	       222.8 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"application/json",_"application/xml"}-8         	 7216270	       171.8 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"application/json",_"application/xml"}-8         	 6906916	       177.6 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"application/json",_"application/xml"}-8         	 7100244	       176.9 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"application/json",_"application/xml"}-8         	 7158033	       169.0 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"application/json",_"application/xml"}-8         	 7086818	       178.0 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"application/json",_"application/xml"}-8         	 7141851	       173.9 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"application/json",_"application/xml"}-8         	 7254651	       170.1 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"application/json",_"application/xml"}-8         	 7252482	       167.7 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"application/json",_"application/xml"}-8         	 7055845	       177.2 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"application/json",_"application/xml"}-8         	 6908438	       170.0 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"application/json",_"application/xml"}-8         	 7294700	       169.2 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"application/json",_"application/xml"}-8         	 6929644	       164.8 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"application/json",_"application/xml"}-8         	 7097929	       178.4 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"application/json",_"application/xml"}-8         	 7053505	       176.6 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_Accepts/run-[]string{"application/json",_"application/xml"}-8         	 6907918	       170.5 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsCharsets-8                                                     	15529420	        76.33 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsCharsets-8                                                     	14623612	        76.62 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsCharsets-8                                                     	15061809	        76.99 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsCharsets-8                                                     	15591081	        77.88 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsCharsets-8                                                     	15324036	        80.08 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsCharsets-8                                                     	15234596	        78.21 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsCharsets-8                                                     	15766075	        76.14 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsCharsets-8                                                     	15616936	        77.10 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsCharsets-8                                                     	15574453	        79.75 ns/op	       0 B/op	       0 allocs/op
Benchmark_Ctx_AcceptsCharsets-8                                                     	15453198	        78.42 ns/op	       0 B/op	       0 allocs/op
…